SL-20014 - vorbis encoding was using random serial numbers
This causes files to be different, even if the audio is the same,
which conflict with content addressing.

This fix causes serial numbers to start at 0 and increase by one
with each block.  This does mean that one cannot concatenate
two vorbis files, but we don't appear to do that anywhere,
so there is that.