DRTVWR-588: Try to make threadsafequeue timing more robust.
The test was coded to push (what's intended to be) the third entry with
timestamp (now + 200ms), then (what's intended to be) the second entry with
timestamp (now + 100ms).

The trouble is that it was re-querying "now" each time. On a slow CI host, the
clock might have advanced by more than 100ms between the first push and the
second -- meaning that the second push would actually have a _later_
timestamp, and thus, even with the queue sorting properly, fail the test's
order validation.

Capture the timestamp once, then add both time deltas to the same time point
to get the relative order right regardless of elapsed real time.